global bpo have a problem

+++ backend/src/main/java/com/example/fifotech/repository/GlobalBpoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GlobalBpoRepository extends JpaRepository<GlobalBPO, Integer> {
+public interface GlobalBpoRepository extends JpaRepository<GlobalBPO, Long> {
 }
